pyproject.toml
[tool.poetry]
name = "crypto-converter"
version = "0.1.0"
description = ""
authors = ["Ivan Sharapenkov <ivansharapenkov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
python-dotenv = "^1.0.1"
websockets = "^12.0"
aiohttp = "^3.10.4"
databases = {extras = ["asyncpg"], version = "^0.9.0"}
fastapi = {extras = ["standard"], version = "^0.112.1"}
uvicorn = {extras = ["standard"], version = "^0.30.6"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.1"
mypy = "^1.11.1"
types-requests = "^2.32.0.20240712"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.select = ["E", "W", "N", "F", "UP", "I", "SIM"]
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"

[tool.ruff.lint.isort]
lines-after-imports = 2
split-on-trailing-comma = false

[tool.mypy]
disallow_untyped_defs = true
disallow_incomplete_defs = true
allow_redefinition = true
python_version = "3.11"
